Replace configure option udevrulesdir with hwdbdir
The --with-udevrulesdir configure option was dropped and a new
--with-hwdbdir option was added instead in upstream 2.0 release.

These values will be picked up by configure through the
udev.pc file and we already build-depend on udev package, so
not really sure why these options are explicitly set.
Maybe only for consistency with --with-systemdsystemunitdir
which is also explicitly specified and needed because we don't
have a build-dependency on systemd (which holds the systemd.pc file).